% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_pallette}
\alias{build_pallette}
\title{Build a gradient pallete based on colors at specific values.}
\usage{
build_pallette(shades, length)
}
\arguments{
\item{shades}{dataframe with columns:
- point - the values at which the pure colors should be set
- color - the pure colors}

\item{length}{number of colors}
}
\description{
Build a gradient pallete based on colors at specific values.
}
