% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{cume_dist_min}
\alias{cume_dist_min}
\title{Return an alternate empirical cumulative distribution, counting only
values with a lower rank.}
\usage{
cume_dist_min(x)
}
\description{
For every value of x, return the number of values with a strictly
lower rank than x.
}

