% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.gpatterns.bam2reads}
\alias{.gpatterns.bam2reads}
\title{bam 2 reads}
\usage{
.gpatterns.bam2reads(
  bams,
  paired_end = TRUE,
  umi1_idx = NULL,
  umi2_idx = NULL,
  add_chr_prefix = FALSE,
  reads_fn = NULL,
  bismark = FALSE,
  bin = system.file("import", "tidy_cpgs.py", package = "gpatterns")
)
}
\description{
bam 2 reads
}
