% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.gpatterns.ds_stats}
\alias{.gpatterns.ds_stats}
\title{ds stats
downsampling stats}
\usage{
.gpatterns.ds_stats(track, dsns, bam = NULL, fastq = NULL,
  tidy_cpgs_dir = NULL, intervals = NULL, sort_rand = FALSE)
}
\arguments{
\item{sort_rand}{}
}
\value{

}
\description{
ds stats
downsampling stats
}
