% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.gpatterns.filter_ontar_reads}
\alias{.gpatterns.filter_ontar_reads}
\title{filter on target reads
Filters on target reads}
\usage{
.gpatterns.filter_ontar_reads(
  reads,
  regions,
  max_distance = 200,
  return_orig = TRUE
)
}
\description{
filter on target reads
Filters on target reads
}
