% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.gpatterns.reads_per_umi}
\alias{.gpatterns.reads_per_umi}
\title{reads per umi}
\usage{
.gpatterns.reads_per_umi(track, dsn = NULL, regions = NULL)
}
\arguments{
\item{dsn}{}
}
\value{

}
\description{
reads per umi
}
