% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils.R
\name{gbin_intervals}
\alias{gbin_intervals}
\title{split genomic intervals to bins}
\usage{
gbin_intervals(intervals, nbins)
}
\arguments{
\item{intervals}{intervals}

\item{nbins}{number of bins}
}
\value{
intervals set with \code{nbins} rows. if \code{nbins} < number of
intervals the original intervals would be returned
}
\description{
split genomic intervals to bins
}

