% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gcluster.run2}
\alias{gcluster.run2}
\title{Runs R commands on a cluster that supports SGE. see \link[misha]{gcluster.run}}
\usage{
gcluster.run2(..., command_list = NULL, opt.flags = "", max.jobs = 400,
  debug = FALSE, R = paste0(R.home(component = "bin"), "/R"),
  packages = NULL, jobs_title = NULL, job_names = NULL,
  collapse_results = FALSE)
}
\arguments{
\item{...}{R commands}

\item{command_list}{list of strings with R commands}

\item{opt.flags}{optional flags for qsub command}

\item{max.jobs}{maximal number of simultaneously submitted jobs}

\item{debug}{if 'TRUE' additional parameters are printed}

\item{R}{command that launches R}

\item{packages}{packages to load for each command}

\item{jobs_title}{title of job names. names would have the title followed
by a serial number}

\item{job_names}{vector with the names of the jobs}

\item{collapse_results}{collapse return values of the jobs to a data frame.
if not possible - would return the usual list.}
}
\value{
if collapse_results is TRUE: data frame with the results of all jobs (rbinded).
if collapse_results is FALSE returns the same as: \link[misha]{gcluster.run}
}
\description{
Runs R commands on a cluster that supports SGE. see \link[misha]{gcluster.run}
}

