% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gextract.left_join}
\alias{gextract.left_join}
\title{Returns the result of track expressions evaluation for each of the
iterator intervals, and cbinds the intervals (instead of intervalID)}
\usage{
gextract.left_join(expr, intervals = NULL, colnames = NULL,
  iterator = NULL, band = NULL, file = NULL, intervals.set.out = NULL,
  suffix = "1")
}
\arguments{
\item{expr}{track expression}

\item{intervals}{genomic scope for which the function is applied}

\item{colnames}{sets the columns names in the returned value. If 'NULL' names are set to track expression.}

\item{iterator}{track expression iterator. If 'NULL' iterator is determined implicitly based on track expressions.}

\item{band}{track expression band. If 'NULL' no band is used.}

\item{file}{file name where the function result is optionally outputed in tab-delimited format}

\item{intervals.set.out}{intervals set name where the function result is optionally outputed}

\item{suffix}{suffix for conflicting column names}
}
\description{
Returns the result of track expressions evaluation for each of the
iterator intervals, and cbinds the intervals (instead of intervalID)
}
\seealso{
\code{\link{misha::gextract}}
}

