% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gintervals.filter}
\alias{gintervals.filter}
\title{Filter intervals set by distance to another.
A wrapper around gintervals.neighbours and filter(dist <= max_distance)}
\usage{
gintervals.filter(intervals1, intervals2, max_distance = 0, abs_dist = TRUE,
  ...)
}
\arguments{
\item{intervals1}{intervals set}

\item{intervals2}{intervals set by which to filter intervals1}

\item{max_distance}{maximal distance of every interval in intervals1 from intervals2 (defualt 0)}

\item{abs_dist}{take the absolute distance}

\item{...}{additional parameters to gintervals.neighbours1}
}
\description{
Filter intervals set by distance to another.
A wrapper around gintervals.neighbours and filter(dist <= max_distance)
}

