% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gintervals.neighbors1}
\alias{gintervals.neighbors1}
\title{Wrapper around gintervals.neighbours that does not return conflicting colnames}
\usage{
gintervals.neighbors1(intervals1 = NULL, intervals2 = NULL, suffix = "1",
  ...)
}
\arguments{
\item{intervals1}{first intervals set}

\item{intervals2}{second intervals set}

\item{suffix}{suffix for conflicting colnames}

\item{...}{other gintervals.neighbours parameters}
}
\description{
Wrapper around gintervals.neighbours that does not return conflicting colnames
}

