% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.adjust_read_pos}
\alias{gpatterns.adjust_read_pos}
\title{Change tidy_cpgs coordinates to fragments coordinates}
\usage{
gpatterns.adjust_read_pos(calls, frag_intervs, maxdist = 0,
  rm_off_target = TRUE)
}
\arguments{
\item{calls}{tidy_cpgs data frame}

\item{frag_intervs}{intervals set of the fragment}

\item{maxdist}{maximal distance from fragments}

\item{rm_off_target}{if TRUE - remove reads with distance > maxdist from frag_intervs
if FALSE - those reads would be left unchanged}
}
\value{
tidy_cpgs data frame with chrom,start,end coordinates adjusted to fragment intervals
}
\description{
changes the coordinates of tidy_cpgs to the fragment coordinates.
This is necessary to avoid over-counting of reads with coordinates that are
shifted by several bps from the restriction site.
}

