% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bissli2.R
\name{gpatterns.bissli2_build}
\alias{gpatterns.bissli2_build}
\title{Create an bowtie2 index for a bisulfite converted genome}
\usage{
gpatterns.bissli2_build(
  reference,
  idx_base,
  bowtie2_options,
  bowtie2_build_bin = "bowtie2-build",
  bissli2_build_bin = .gpatterns.bissli2_build_bin
)
}
\arguments{
\item{reference}{a vector of FASTA file names holding the reference genome.}

\item{idx_base}{The basename of the index files that will be created.}

\item{bowtie2_options}{Any unknown options are passed transparently to bowtie2-build.
Note that bissli2-build can only handle FASTA files as input. Therefore the '-f' option is always passed to bowtie2-build}

\item{bowtie2_build_bin}{The path of the bowtie2-build executable. [bowtie2-build]}

\item{bissli2_build_bin}{binary of bissli2-build}
}
\value{

}
\description{
Create an bowtie2 index for a bisulfite converted genome
}
