% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipoly.R
\name{gpatterns.border_pattern}
\alias{gpatterns.border_pattern}
\title{Returns TRUE if pattern is a 'border' pattern}
\usage{
gpatterns.border_pattern(pattern)
}
\arguments{
\item{pattern}{pattern}
}
\description{
Returns TRUE if pattern is a 'border' pattern
}

