% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipoly.R
\name{gpatterns.calc_bipolarity}
\alias{gpatterns.calc_bipolarity}
\title{Run bipolar model}
\usage{
gpatterns.calc_bipolarity(track, uniform_mix = 0.05,
  max_sampling_n = 1000, min_sampling_n = 100, init_num = 3,
  min_pat_cov = 5, save_tab = FALSE, overwrite = FALSE,
  parallel = getOption("gpatterns.parallel"),
  thread_num = getOption("gpatterns.parallel.thread_num"),
  verbose = FALSE, use_sge = FALSE, max_jobs = 300, ...)
}
\arguments{
\item{track}{track}

\item{uniform_mix}{uniform_mix}

\item{max_sampling_n}{max_sampling_n}

\item{min_sampling_n}{min_sampling_n}

\item{init_num}{init_num}

\item{min_pat_cov}{min_pat_cov}

\item{save_tab}{save_tab}

\item{overwrite}{overwrite}

\item{parallel}{parallel}

\item{thread_num}{thread_num}

\item{verbose}{verbose}

\item{use_sge}{use_sge}

\item{max_jobs}{max_jobs}

\item{...}{...}
}
\description{
Run bipolar model
}
