% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipoly.R
\name{gpatterns.calc_epipoly}
\alias{gpatterns.calc_epipoly}
\title{Calculate epipolymorphism of samples}
\usage{
gpatterns.calc_epipoly(patterns)
}
\arguments{
\item{patterns}{patterns}
}
\description{
Calculate epipolymorphism of samples
}
