% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{gpatterns.chip2peaks}
\alias{gpatterns.chip2peaks}
\title{Calculate peaks of chip-seq tracks}
\usage{
gpatterns.chip2peaks(chip_tracks, quant_thresh = 0.999,
  normalize = NULL, logical_gate = "|")
}
\arguments{
\item{chip_tracks}{names of chip-seq tracks}

\item{quant_thresh}{quantile of chip signal that would be considered as peak}

\item{normalize}{normalize all the regions to have the same size.
if NULL no normalization would be done.}

\item{logical_gate}{logical function to combine the different tracks, e.g. '|' or '&'}
}
\value{
intervals set with peaks
}
\description{
Calculate peaks of chip-seq tracks
}
