% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipoly.R
\name{gpatterns.clean_pattern}
\alias{gpatterns.clean_pattern}
\title{Returns TRUE if pattern is fully methylated / unmethylated and FALSE if not}
\usage{
gpatterns.clean_pattern(pattern)
}
\arguments{
\item{pattern}{pattern}
}
\description{
Returns TRUE if pattern is fully methylated / unmethylated and FALSE if not
}

