% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.cluster_avg_meth}
\alias{gpatterns.cluster_avg_meth}
\title{Cluster average methylation of multiple tracks}
\usage{
gpatterns.cluster_avg_meth(avgs, K, verbose = FALSE,
  clust_order_func = partial(mean, na.rm = T), clust_columns = FALSE,
  column_k = NULL, ret_hclust = FALSE, tidy = TRUE,
  intra_clust_order = TRUE)
}
\arguments{
\item{avgs}{avgs pre-computed intervals set with average methylation, output of gpatterns.get_avg_meth.}

\item{K}{number of clusters for kmeans++}

\item{verbose}{display kmeans++ messages}

\item{clust_order_func}{order clusters by clust_order_func [default: mean]}

\item{clust_columns}{cluster the columns (samples) using hclust}

\item{column_k}{cut the hclust of the columns and return an additional
'samp_clust' column with the column cluster}

\item{ret_hclust}{return a list with avgs (the clustering result) and
hc (the hclust object)}

\item{tidy}{is the input tidy}

\item{intra_clust_order}{order within each cluster using hclust (otherwise, the order would be by genomic position)}
}
\value{
if tidy: data frame with the following fields:
\itemize{
     \item{'chrom'}{}
     \item{'start'}{}
     \item{'end'}{}
     \item{'ord'}{order of intervals within the cluster}
     \item{'samp'}{sample id}
     \item{'samp_clust'}{id of the column cluster (hclust cutree)}
     \item{'clust'}{cluster id}
     \item{'original columns'}{original columns of avgs (meth,unmeth,avg,cov)}
}
if not tidy: data frame with the following fields:
\itemize{
     \item{'chrom'}{}
     \item{'start'}{}
     \item{'end'}{}
     \item{'ord'}{order of intervals within the cluster}
     \item{'samp_clust'}{id of the column cluster (hclust cutree)}
     \item{'clust'}{cluster id}
     \item{'original columns'}{original columns of avgs}
}
}
\description{
Cluster average methylation of multiple tracks
}
