% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.count_cpgs}
\alias{gpatterns.count_cpgs}
\title{count CpGs in intervals}
\usage{
gpatterns.count_cpgs(intervals, cgs_track = .gpatterns.genome_cpgs_track)
}
\arguments{
\item{intervals}{intervals to filter}

\item{cgs_track}{cpgs track (has '1' at every CpG in the genome and '0' or NA otherwise)}
}
\description{
count CpGs in intervals
}

