% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.create_patterns_track}
\alias{gpatterns.create_patterns_track}
\title{Creates patterns attributes of track}
\usage{
gpatterns.create_patterns_track(track, description, pat_space,
  patterns_tab = NULL, add_read_id = TRUE, noise_threshold = 0.2,
  overwrite = TRUE, canonize = FALSE, add_biploar_stats = FALSE, ...)
}
\arguments{
\item{track}{track name}

\item{description}{a character string description}

\item{pat_space}{pattern space data frame
(output of gpatterns.tracks_to_pat_space or gpatterns.intervs_to_pat_space)
needs to have the following fields: chrom,start,end,fid}

\item{patterns_tab}{table with patterns (output of gpatterns.tidy_cpgs_to_pats)
needs to have the following fields: fid,pattern}

\item{add_read_id}{save the read_id together with the patterns}

\item{noise_threshold}{threshold to consider pattern as 'noise'}

\item{overwrite}{overwrite existing tracks}

\item{canonize}{convert to canonic form (see: \code{\link[gintervals.canonic]{misha}}).
warning: may lose data}

\item{add_biploar_stats}{run mixture model and add bipolaritly stats.
(warning: heavy computation)}

\item{...}{additional parameters for gpatterns.calc_bipolarity}
}
\description{
Creates patterns attributes of track
}

