% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.downsample_patterns}
\alias{gpatterns.downsample_patterns}
\title{Downsample patterns table}
\usage{
gpatterns.downsample_patterns(patterns, dsn)
}
\arguments{
\item{patterns}{patterns table}

\item{dsn}{downsampling n}
}
\description{
Downsample patterns table
}

