% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipoly.R
\name{gpatterns.epipoly_plot}
\alias{gpatterns.epipoly_plot}
\title{Plots epipolymorphism}
\usage{
gpatterns.epipoly_plot(tracks, intervals, colnames = NULL, pat_len = 5,
  median_line_binsize = 6, meth_binsize = 0.01, fig_ofn = NULL,
  width = NULL, height = NULL, ncolumns = 3, separate_figs = FALSE)
}
\arguments{
\item{tracks}{track names}

\item{intervals}{genomic scope for which the function is applied}

\item{colnames}{column names for the samples (useful when track names are too long}

\item{pat_len}{pattern length}

\item{median_line_binsize}{number of bins to smooth in median line}

\item{meth_binsize}{size of bin for methyltion in x axis}

\item{fig_ofn}{figure file name. if NULL plotting would occur on the current device}

\item{width}{figure width in inches}

\item{height}{figure width in inches}

\item{ncolumns}{number of columns}

\item{separate_figs}{plot each track in a separate figure}
}
\value{
plot. If seperate.figs=TRUE list of plots
}
\description{
Plots epipolymorphism
}
\examples{

}
