% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.extract}
\alias{gpatterns.extract}
\title{Extract pattern data}
\usage{
gpatterns.extract(..., intervals, colnames = NULL, iterator = NULL,
  cpg.nhood = 500, elements = c("cpg.content", "fid", "ncpg", "n", "n0",
  "n1", "nx", "meth"), add_bipolar_stats = FALSE, tidy = TRUE)
}
\arguments{
\item{...}{track names}

\item{intervals}{intervals}

\item{colnames}{colnames}

\item{iterator}{iterator}

\item{cpg.nhood}{cpg.nhood}

\item{elements}{elements}

\item{add_bipolar_stats}{add_bipolar_stats}

\item{tidy}{tidy}
}
\description{
Extract pattern data
}

