% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.extract_all}
\alias{gpatterns.extract_all}
\title{Extract pattern data from tracks without specifying coordinates}
\usage{
gpatterns.extract_all(
  ...,
  samples = NULL,
  elements = c("fid", "ncpg", "n", "n0", "n1", "nx", "pat_meth", "epipoly"),
  add_bipolar_stats = FALSE,
  dsn = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{...}{tracks (comma separated)}

\item{samples}{name for each track}

\item{elements}{elements to extract}

\item{add_bipolar_stats}{add bipolar stats}

\item{na.rm}{remove rows with missing data}

\item{extract}{downsampled data}
}
\value{

}
\description{
Extract pattern data from tracks without specifying coordinates
}
