% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.extract_patterns}
\alias{gpatterns.extract_patterns}
\title{Extract patterns of a track}
\usage{
gpatterns.extract_patterns(track, fids = NULL, tidy = TRUE, dsn = NULL)
}
\arguments{
\item{track}{name of track}

\item{fids}{extract for specific fragments}

\item{tidy}{if TRUE: returns tidy format, else - a list}

\item{dsn}{extract downsampled patterns}
}
\description{
Extract patterns of a track
}

