% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.filter_by_dist}
\alias{gpatterns.filter_by_dist}
\title{Filter intervals set by distance from another intervals set}
\usage{
gpatterns.filter_by_dist(
  intervals,
  annot_intervs = NULL,
  annot_intervs_dist = 0,
  orientation = "upstream",
  use_tss = FALSE,
  tss_max_dist = 1000
)
}
\arguments{
\item{intervals}{intervals to filter}

\item{annot_intervs}{annotation intervals to filter by}

\item{annot_intervs_dist}{maximal distance from annot_intervs}

\item{orientation}{orientation (upstream, downstream or both)}

\item{use_tss}{use TSS as annot intervs}

\item{tss_dist}{annot_intervs_dist for TSS}
}
\value{

}
\description{
Filter intervals set by distance from another intervals set
}
