% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.filter_cpgs}
\alias{gpatterns.filter_cpgs}
\title{Filter intervals by number of CpGs}
\usage{
gpatterns.filter_cpgs(intervals, min_cpgs,
  cgs_track = .gpatterns.genome_cpgs_track)
}
\arguments{
\item{intervals}{intervals to filter}

\item{min_cpgs}{minimal number of CpGs}

\item{cgs_track}{cpgs track (has '1' at every CpG in the genome and '0' otherwise)}
}
\description{
Filter intervals by number of CpGs
}
