% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gpatterns.filter_loci}
\alias{gpatterns.filter_loci}
\title{Filter loci with low variance across samples}
\usage{
gpatterns.filter_loci(
  avgs,
  k = NULL,
  center_sd_thresh = NULL,
  min_loci_frac = 0.3,
  seed = NULL,
  tidy = TRUE,
  avg_col = "avg",
  plot_cluster_sd = FALSE,
  ret_clust_sd = FALSE
)
}
\arguments{
\item{avgs}{'tidy' output of \code{gpatterns.get_avg_meth}. In order to filter using the regularized (imputed) values (output of \code{gpatterns.impute}), set \code{avg_col} to 'avg_reg'}

\item{k}{number of clusters to divide to. if NULL - k would be chosen as number of loci divided by 150, 
in order to have ~100 loci per cluster.}

\item{center_sd_thresh}{minimal sd of the cluster centers. if NULL - \code{center_sd_thresh} would be set by taking
the sd of the cluster with the largest difference to the next cluster sd that still
leaves \code{min_loci_frac} loci.}

\item{min_loci_frac}{minimal fraction of loci to return in case \code{center_sd_thesh == NULL}.}

\item{seed}{seed to use in \code{TGL_kmeans_tidy}}

\item{tidy}{return tidy output}

\item{avg_col}{column of average methylation in \code{avgs}.}

\item{plot_cluster_sd}{plot the standard deviation of clusters.}

\item{ret_clust_sd}{return a list with the filtered loci (under 'avgs'), and the cluster sd plot (under 'clust_sd_p')}
}
\value{
see \code{ret_clust_sd}. if \code{tidy}: \code{avgs} data frame without the filtered loci. if \code{tidy == FALSE}: intervals set with additional fields with regularized average methylation for each sample.
}
\description{
Clusters the loci with kmeans with a relatively high k, calculates the \code{sd} of each cluster centers, 
and then removes loci that are within clusters with low variance.
}
