% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.frag_stats}
\alias{gpatterns.frag_stats}
\title{transforms patterns to summary statistics: n,n0,n1,nx,nc,meth and epipoly}
\usage{
gpatterns.frag_stats(patterns_tab, noise_threshold = 0.2)
}
\arguments{
\item{patterns_tab}{patterns table (needs to have 'fid' and 'pattern' fields)}

\item{noise_threshold}{thershold of a pattern to be considered 'noise'}
}
\value{

}
\description{
transforms patterns to summary statistics: n,n0,n1,nx,nc,meth and epipoly
}
