% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.get_avg_meth}
\alias{gpatterns.get_avg_meth}
\title{Extract average methylation}
\usage{
gpatterns.get_avg_meth(tracks, intervals, iterator = NULL, min_cov = NULL,
  mask_by_cov = FALSE, use_cpgs = FALSE, min_samples = 1,
  min_cpgs = NULL, min_var = NULL, var_quantile = NULL, names = NULL,
  tidy = TRUE)
}
\arguments{
\item{tracks}{methylation tracks}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{see iterator in \code{\link[gextract]{misha}}. if NULL iterator
would be set to CpGs}

\item{min_cov}{minimal coverage for iterator interval}

\item{mask_by_cov}{change loci with coverage < min_cov to NA}

\item{use_cpgs}{use CpGs as iterator}

\item{min_samples}{minimal number of samples with cov >= min_cov. if min_cov
is NULL it would be set to 1.}

\item{min_cpgs}{minimal number of CpGs per iterator interval}

\item{min_var}{minimal variance (across samples) per iterator interval}

\item{var_quantile}{minimal quantile of variance per iterator interval}

\item{names}{alternative names to tracks. similar to colnames in
\code{\link[gextract]{misha}} if tidy == FALSE.}

\item{tidy}{if TRUE returns a tidy data frame with the following fields:
chrom, start, end, intervalID, samp, meth, unmeth, avg, cov.
if FALSE returns a data frame with average methylation,
similar to \code{\link[gextract]{misha}}'}
}
\description{
Extract average methylation
}

