% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.get_fragment_cov}
\alias{gpatterns.get_fragment_cov}
\title{Get fragment coverage}
\usage{
gpatterns.get_fragment_cov(track, intervals = NULL, save_track = FALSE,
  parallel = getOption("gpatterns.parallel"), ...)
}
\arguments{
\item{track}{track}

\item{intervals}{fragment intervals. if NULL - fragments would be based chrom,start,end positions
of tidy_cpgs, assuming gpatterns.adjust_read_pos was called before (as in any
run of gpatterns.import_* with frag_intervs parameter).}

\item{parallel}{parallel}

\item{...}{other parameters of gpatterns.adjust_read_pos}
}
\value{
data frame with fragmets coordinates (chrom,start,end) and 'frag_cov'
field with the fragment coverage
}
\description{
get number of molecules per each fragment for a track. This can
give us some upper bound of the number of methylation patterns per fragment.
}
