% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.get_pat_cov}
\alias{gpatterns.get_pat_cov}
\title{Calculate pattern coverage for each CpG}
\usage{
gpatterns.get_pat_cov(track, pat_len, max_span = 500, by_coord_bin = TRUE,
  parallel = getOption("gpatterns.parallel"))
}
\arguments{
\item{track}{name of track}

\item{pat_len}{pattern length}

\item{max_span}{maximum span to look for CpGs}

\item{by_coord_bin}{execute separatly for each coordinates bin
(for tracks that do not fit into memory)}

\item{parallel}{execute parallely}
}
\description{
Calculate pattern coverage for each CpG
}

