% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gpatterns.get_pipeline_stats}
\alias{gpatterns.get_pipeline_stats}
\title{Get QC statistics for track}
\usage{
gpatterns.get_pipeline_stats(track, tidy_cpgs_stats_dir,
  uniq_tidy_cpgs_stats_dir, add_mapping_stats = FALSE)
}
\arguments{
\item{track}{track name}

\item{tidy_cpgs_stats_dir}{directory with tidy_cpgs stats files}

\item{uniq_tidy_cpgs_stats_dir}{directory with filter_dups stats files}

\item{add_mapping_stats}{add full mapping statistics (singles, discordant etc.)}
}
\description{
Get QC statistics for track
}

