% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.get_promoters}
\alias{gpatterns.get_promoters}
\title{Define promoter regions}
\usage{
gpatterns.get_promoters(upstream = 500, downstream = 50)
}
\arguments{
\item{upstream}{bp upstream to TSS}

\item{downstream}{bp downstread from tss}
}
\description{
Define promoter regions
}
