% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.get_tidy_cpgs}
\alias{gpatterns.get_tidy_cpgs}
\title{Get tidy cpgs from a track}
\usage{
gpatterns.get_tidy_cpgs(track, intervals = NULL)
}
\arguments{
\item{track}{name of track}

\item{intervals}{intervals}
}
\description{
Get tidy cpgs from a track
}

