% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gpatterns.impute}
\alias{gpatterns.impute}
\title{Impute missing methylation calls}
\usage{
gpatterns.impute(avgs, k_reg = 3, min_locus_cov = NULL, tidy = TRUE)
}
\arguments{
\item{avgs}{'tidy' output of \code{gpatterns.get_avg_meth}.}

\item{k_reg}{k_reg}

\item{min_locus_cov}{minimal total coverage per locus. If not NULL, filters out loci with less
than \code{min_locus_cov} methylation calls.}

\item{tidy}{return 'tidy' output.}
}
\value{
if \code{tidy}, \code{avgs} with additional \code{meth_reg}, \code{cov_reg} and \code{avg_reg} fields with the imputed methylation calls, total number of calls and avergae methylation (respectively) for each sample.
If \code{tidy == FALSE}: intervals set with additional fields with regularized average methylation for each sample.
}
\description{
Impute missing methylation calls
}
