% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.intervals_coverage}
\alias{gpatterns.intervals_coverage}
\title{Get coverage of intervals in multiple tracks}
\usage{
gpatterns.intervals_coverage(
  tracks,
  intervals,
  iterator = NULL,
  file = NULL,
  intervals.set.out = NULL,
  pre_screen = TRUE,
  frag_cov = FALSE
)
}
\arguments{
\item{tracks}{methylation tracks}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{see iterator in \link[misha]{gextract}. if NULL iterator
would be set to CpGs}

\item{file}{save output to file (only in non tidy mode, would not filer by variance)}

\item{intervals.set.out}{save output big intervals set (only in tidy mode,
would not filter by variance)}

\item{pre_screen}{screen for covered intervals before calculating the total coverage (for large intervals set)}

\item{frag_cov}{use fragment coverage (if exists)}
}
\value{
intervals set with 'sum' column containing the total coverage (sum) in all the tracks
}
\description{
Get coverage of intervals in multiple tracks
}
