% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gpatterns.ls}
\alias{gpatterns.ls}
\title{Returns a list of track names}
\usage{
gpatterns.ls(
  regex = "",
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{regex}{regular expression of track names}

\item{ignore.case}{see 'grep'}

\item{perl}{see 'grep'}

\item{fixed}{see 'grep'}

\item{useBytes}{see 'grep'}
}
\value{
An array that contains the names of tracks that match the supplied patterns
}
\description{
Returns a list of gpatterns tracks in the genomic database
}
