% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.merge_tracks}
\alias{gpatterns.merge_tracks}
\title{Merge tracks}
\usage{
gpatterns.merge_tracks(tracks, new_track, description, iterator = NULL,
  intervals = gintervals.all(), add_var = FALSE)
}
\arguments{
\item{tracks}{tracks to merge}

\item{new_track}{new track name}

\item{description}{new track description}

\item{iterator}{new track iterator}

\item{intervals}{intervals scope}

\item{add_var}{calculate variance (of tracks '.avg'). Will be created in new_track.var}
}
\value{
name of the new_track
}
\description{
Merge tracks
}

