% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{gpatterns.pipeline}
\alias{gpatterns.pipeline}
\title{Run pipeline}
\usage{
gpatterns.pipeline(
  config_file,
  log_file = NULL,
  config_dir = NULL,
  run_commands = TRUE,
  run_dir = getwd(),
  overwrite = FALSE
)
}
\description{
Run pipeline
}
