% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.plot_cg_cor}
\alias{gpatterns.plot_cg_cor}
\title{Plot CpG correlation (pearson) as a function of distance between CpGs}
\usage{
gpatterns.plot_cg_cor(tracks, dist_breaks, intervals = gintervals.all(),
  adjacent = FALSE, names = NULL, width = 500, height = 260,
  fig_fn = NULL, xlab = "Distance between CpGs",
  ylab = "CpG correlation", colors = NULL,
  parallel = getOption("gpatterns.parallel"), ...)
}
\arguments{
\item{tracks}{tracks to plot}

\item{dist_breaks}{distance breaks}

\item{intervals}{genomic scope for which the function is applied}

\item{adjacent}{plot correlation only between adjacent CpGs (faster)}

\item{names}{alternative names for the track}

\item{width}{plot width (if fig_fn is not NULL)}

\item{height}{plot height (if fig_fn is not NULL)}

\item{fig_fn}{output filename for the figure (if NULL, figure would be returned)}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{colors}{custom colors}

\item{parallel}{parallel}

\item{...}{}
}
\value{
list with trend data frame (under 'trend') and the plot (under 'p')
}
\description{
Plot CpG correlation (pearson) as a function of distance between CpGs
}
