% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{gpatterns.plot_chip_quantiles}
\alias{gpatterns.plot_chip_quantiles}
\title{Plot chip-seq singal with quantile lines}
\usage{
gpatterns.plot_chip_quantiles(
  chip_track,
  intervals,
  quantiles = c(0.8, 0.9, 0.95, 0.99, 0.995, 0.999)
)
}
\arguments{
\item{chip_track}{chip-seq track}

\item{intervals}{intervals to plot at}

\item{quantiles}{line quantiles to plot}
}
\description{
Plot chip-seq singal with quantile lines
}
