% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.plot_clustering}
\alias{gpatterns.plot_clustering}
\title{Plot a heatmap of average methylation clustering}
\usage{
gpatterns.plot_clustering(avgs, rows_clust_method = "kmeans", K = NULL,
  cluster_cols = TRUE, show_row_clust = TRUE, annotation = NULL,
  annotation_colors = NULL, annotation_col = NULL, show_rownames = FALSE,
  show_colnames = FALSE, cluster_gaps = FALSE, color_pal = .blue_red_pal,
  fontsize = 8, fontsize_row = fontsize, fontsize_col = fontsize,
  main = NULL, device = "png", fig_ofn = NULL, width = 2100,
  height = 2000, res = 300, plot = TRUE, motif_enrich = NULL,
  cluster_motifs = TRUE, qval_thresh = 0.05, show_sig = TRUE,
  motif_fontsize_col = fontsize * 0.8, motif_fontsize_sig = fontsize * 1.3,
  ...)
}
\arguments{
\item{avgs}{output of gpatterns.cluster_avg_meth}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{annotation}{deprecated parameter that currently sets the annotation_col if it is missing}

\item{annotation_colors}{list for specifying annotation_row and
annotation_col track colors manually. It is  possible to define the colors
for only some of the features. Check examples for  details.}

\item{annotation_col}{similar to annotation_row, but for columns.}

\item{show_rownames}{boolean specifying if column names are be shown.}

\item{show_colnames}{boolean specifying if column names are be shown.}

\item{fontsize}{base fontsize for the plot}

\item{fontsize_row}{fontsize for rownames (Default: fontsize)}

\item{fontsize_col}{fontsize for colnames (Default: fontsize)}

\item{main}{the title of the plot}

\item{fig_ofn}{output filename of the figure. if NULL would plot to current device}

\item{width}{width ('png' parameter)}

\item{height}{height ('png' parameter)}

\item{plot}{if FALSE, would only return the clustering}

\item{...}{other pheatmap1 parameters}
}
\description{
Plot a heatmap of average methylation clustering
}
