% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{gpatterns.putative_enhancers}
\alias{gpatterns.putative_enhancers}
\title{Define putative enhancer regions from chip-seq tracks}
\usage{
gpatterns.putative_enhancers(chip_tracks, quant_thresh = 0.999,
  normalize = NULL, min_tss_dist = 2000, logical_gate = "|")
}
\arguments{
\item{chip_tracks}{names of chip-seq tracks}

\item{quant_thresh}{quantile of chip signal that would be considered as peak}

\item{normalize}{normalize all the regions to have the same size.
if NULL no normalization would be done.}

\item{min_tss_dist}{minimal (absolute) distance from TSS}

\item{logical_gate}{logical function to combine the different tracks, e.g. '|' or '&'}
}
\value{
intervals set with putative enhancers
}
\description{
Define putative enhancer regions from chip-seq tracks
}

