% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.screen_by_coverage}
\alias{gpatterns.screen_by_coverage}
\title{screen intervals by coverage in multiple tracks}
\usage{
gpatterns.screen_by_coverage(tracks, intervals, iterator, min_cov,
  min_samples = NULL, intervals.set.out = NULL)
}
\arguments{
\item{tracks}{methylation tracks}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{see iterator in \code{\link[gextract]{misha}}. if NULL iterator
would be set to CpGs}

\item{min_cov}{minimal coverage for iterator interval}

\item{min_samples}{minimal number of samples with cov >= min_cov. if min_cov
is NULL it would be set to 1.}

\item{intervals.set.out}{intervals.set.out}
}
\description{
screen intervals by coverage in multiple tracks
}

