% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.separate_strands}
\alias{gpatterns.separate_strands}
\title{Separate a track by strands (for QC purposes)}
\usage{
gpatterns.separate_strands(track, description, out_track = NULL,
  intervals = NULL, minus_suffix = ".minus", plus_suffix = ".plus", ...)
}
\arguments{
\item{out_track}{output track name if NULL \code{track} would be used}

\item{intervals}{intervals to extract from tidy_cpgs}

\item{minus_suffix}{suffix for the minus strand track}

\item{plus_suffix}{suffix for the plus strand track}

\item{...}{additional parameters for \code{gpatterns.import_from_tidy_cpgs}}
}
\value{
none
}
\description{
Separate a track by strands (for QC purposes)
}
