% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.smoothScatter}
\alias{gpatterns.smoothScatter}
\title{Plots smoothScatter of average methylation of two tracks}
\usage{
gpatterns.smoothScatter(samples, avgs = NULL, fig_ofn = NULL,
  sample_names = NULL, width = 300, height = 300,
  color.pal = .smooth_scatter_pal2, device = "png", title_text = NULL,
  add_n = TRUE, min_cov = NULL, min_samples = 2, iterator = NULL, ...)
}
\arguments{
\item{samples}{a vector of 2 sample names as present in 'samp' filed of avgs, or track names if
avgs is NULL}

\item{avgs}{pre-computed intervals set with average methylation, output of gpatterns.get_avg_meth.
if NULL, gpatterns.get_avg_meth(tracks=samples, ...) would be called. Note that
you using this option together with additional parameters to smoothScatter would
produce warnings.}

\item{fig_ofn}{figure filename. if NULL figure would be plotted on the current device}

\item{sample_names}{names for the two tracks to plot as the axis labels.
if NULL the sample names would be used.}

\item{width}{plot width}

\item{height}{plot height}

\item{color.pal}{color pallete}

\item{device}{function to open a device if fig_ofn is not NULL. default: 'png'}

\item{title_text}{title text}

\item{add_n}{add 'n = number_of_observations' to the title}

\item{min_cov}{minimal coverage for iterator interval.}

\item{...}{other parameters for smoothScatter / gpatterns.get_avg_meth.}
}
\value{
None
}
\description{
Plots smoothScatter of average methylation of two tracks
}
\examples{

}
