% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.spatial_meth_trend}
\alias{gpatterns.spatial_meth_trend}
\title{Plot spatial trend of methylation around intervals}
\usage{
gpatterns.spatial_meth_trend(tracks, intervals, method = "extract",
  dist_breaks = seq(-1000, 1000, 200),
  iterator = .gpatterns.genome_cpgs_intervals, min_cov = NULL,
  min_cgs = NULL, names = NULL, width = 500, height = 260,
  fig_fn = NULL, xlab = "Distance (bp)", title = "", legend = TRUE,
  colors = NULL, parallel = getOption("gpatterns.parallel"))
}
\arguments{
\item{tracks}{tracks to plot}

\item{intervals}{intervals to plot spatial methylation around.
Can be any intervals set, or one of the following 'special' intervals:
'tss', 'exon', 'utr3', 'intron', 'cgi'}

\item{dist_breaks}{breaks to determine the distances from intervals}

\item{iterator}{track expression iterator (of both tracks and strat_track)}

\item{min_cov}{minimal coverage of each track}

\item{min_cgs}{minimal number of CpGs per bin.}

\item{names}{alternative names for the track}

\item{width}{plot width (if fig_fn is not NULL)}

\item{height}{plot height (if fig_fn is not NULL)}

\item{fig_fn}{output filename for the figure (if NULL, figure would be returned)}

\item{xlab}{label for the x axis'}

\item{title}{title for the plot}

\item{legend}{add legend}

\item{colors}{custom colors}

\item{parallel}{get trends parallely}
}
\value{
list with trend data frame (under 'trend') and the plot (under 'p')
}
\description{
calculates the distance of \code{tracks} values from \code{intervals}
bins the distances based on \code{dist_breaks}, and then calculates the average
methylation \code{m / (m + um)} in each bin and plots the results.
}

