% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.tidy_cpgs_2_pat_freq}
\alias{gpatterns.tidy_cpgs_2_pat_freq}
\title{Generate CpG pattern frequency from tidy_cpgs}
\usage{
gpatterns.tidy_cpgs_2_pat_freq(calls, pat_length = 2, min_cov = 1,
  tidy = TRUE)
}
\arguments{
\item{calls}{tidy_cpgs}

\item{pat_length}{length of the pattern, e.g. for pat_lenth == 2 patterns would
be 00,01,10,11}

\item{min_cov}{minimal coverage}

\item{tidy}{tidy}
}
\description{
Generate CpG pattern frequency from tidy_cpgs
}

