% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.tidy_cpgs_2_pileup}
\alias{gpatterns.tidy_cpgs_2_pileup}
\title{Generate pileup form tidy_cpgs}
\usage{
gpatterns.tidy_cpgs_2_pileup(calls, dsn = NULL)
}
\arguments{
\item{calls}{tidy_cpgs}

\item{dsn}{downsampling n}
}
\value{
data frame with chrom,start,end,meth,unmeth,cov,avg
}
\description{
Generate pileup form tidy_cpgs
}
