% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.tidy_cpgs_to_pats}
\alias{gpatterns.tidy_cpgs_to_pats}
\title{Extracts patterns given pattern space}
\usage{
gpatterns.tidy_cpgs_to_pats(track, pattern_space = NULL,
  pattern_space_file = NULL, max_missing = Inf, min_cov = 1,
  by_coord_bin = TRUE, parallel = getOption("gpatterns.parallel"), ...)
}
\arguments{
\item{track}{name of track}

\item{pattern_space}{pattern space data frame
(output of gpatterns.tracks_to_pat_space or gpatterns.intervs_to_pat_space)
needs to have the following fields: chrom,start,end,fid}

\item{pattern_space_file}{file that holds the pattern space}

\item{max_missing}{maximal number of missing data ('*') in pattern}

\item{min_cov}{minimal pattern coverage}

\item{by_coord_bin}{execute separatly for each coordinates bin
(for tracks that do not fit into memory)}

\item{...}{additional parameters to gpatterns.apply_tidy_cpgs}
}
\description{
Extracts patterns given pattern space
}
