% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.track_pat_space}
\alias{gpatterns.track_pat_space}
\title{Get pattern space of a track}
\usage{
gpatterns.track_pat_space(track)
}
\arguments{
\item{track}{track name}
}
\value{
pattern space of a track
}
\description{
Get pattern space of a track
}
