% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.tracks_to_pat_space}
\alias{gpatterns.tracks_to_pat_space}
\title{Generate pattern space}
\usage{
gpatterns.tracks_to_pat_space(tracks, pat_len, max_span = NULL,
  exclude_intervals = NULL, exclude_cpgs = NULL, min_cov = 1)
}
\arguments{
\item{tracks}{tracks}

\item{pat_len}{pattern length}

\item{max_span}{maximum span to look for CpGs}

\item{exclude_intervals}{intervals to exclude}

\item{exclude_cpgs}{CpGs to exclude}

\item{min_cov}{minimal coverage per CpG}
}
\description{
Generate pattern space
}
