% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils.R
\name{gtrack.array.import_from_df}
\alias{gtrack.array.import_from_df}
\title{Import array from data frame}
\usage{
gtrack.array.import_from_df(df, track, description)
}
\arguments{
\item{df}{data frame with intervals and values}

\item{track}{name of the newly created track}

\item{description}{a character string description}
}
\description{
Import array from data frame
}
