% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gtrack.create_sparse}
\alias{gtrack.create_sparse}
\title{overrides gtrack.create_sparse without rescaning database}
\usage{
gtrack.create_sparse(track = NULL, description = NULL, intervals = NULL,
  values = NULL, rescan = FALSE)
}
\description{
overrides gtrack.create_sparse without rescaning database
}

