% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils.R
\name{gtrack.import_mappedseq_bam}
\alias{gtrack.import_mappedseq_bam}
\title{Wrapper around gtrack.import_mappedseq for (multiple) bam files}
\usage{
gtrack.import_mappedseq_bam(bam_files, ...)
}
\arguments{
\item{bam_files}{vector of bam (or sam) files to import}

\item{...}{parameters of gtrack.import_mappedseq}
}
\value{

}
\description{
Wrapper around gtrack.import_mappedseq for (multiple) bam files
}
\seealso{
\link[misha]{gtrack.import_mappedseq}
}
