% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gvextract}
\alias{gvextract}
\title{Creates a virtual track and runs gextract}
\usage{
gvextract(tracks, intervals, colnames = NULL, iterator = NULL,
  band = NULL, file = NULL, intervals.set.out = NULL, func = NULL,
  params = NULL)
}
\arguments{
\item{params}{}
}
\description{
Creates a virtual track and runs gextract
}

