% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_smoothScatter}
\alias{p_smoothScatter}
\title{Smoothscatter that works with the pipe operator}
\usage{
p_smoothScatter(.data, .x, .y, xlab = NULL, ylab = NULL, ...)
}
\description{
Smoothscatter that works with the pipe operator
}
