% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGL_kmeans.R
\name{TGL_kmeans}
\alias{TGL_kmeans}
\title{Effiecient implementation of kmeans++ algorithm}
\usage{
TGL_kmeans(tab, K, id_column = FALSE, allow_nas = TRUE, verbose = FALSE,
  fn = NULL, tab_file = FALSE, TGL.kmeans.bin = sprintf("\%s/\%s",
  system.file("bin", package = "gpatterns"), "TGLKMeans_static"))
}
\arguments{
\item{tab}{table}

\item{K}{K}

\item{id_column}{id_column}

\item{allow_nas}{allow_nas}

\item{verbose}{verbose}

\item{fn}{fn}

\item{tab_file}{tab_file}

\item{TGL.kmeans.bin}{TGL.kmeans.bin}
}
\description{
Effiecient implementation of kmeans++ algorithm
}

