% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGL_kmeans.R
\name{TGL_kmeans2}
\alias{TGL_kmeans2}
\title{Alternative interface for TGL_kmeans
Returned value is more similar to native R kmeans
Allows keeping the output of the kmeans exec within the R return value}
\usage{
TGL_kmeans2(data, k, keep.log = FALSE, tmpdir = NULL, tmpfn = NULL,
  TGL.kmeans.bin = sprintf("\%s/\%s", system.file("bin", package =
  "gpatterns"), "TGLKMeans_static"))
}
\arguments{
\item{TGL.kmeans.bin}{}
}
\description{
Alternative interface for TGL_kmeans
Returned value is more similar to native R kmeans
Allows keeping the output of the kmeans exec within the R return value
}
