% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils.R
\name{gcluster.run2}
\alias{gcluster.run2}
\title{Runs R commands on a cluster that supports SGE}
\usage{
gcluster.run2(
  ...,
  command_list = NULL,
  opt.flags = "",
  max.jobs = 400,
  debug = FALSE,
  R = paste0(R.home(component = "bin"), "/R"),
  packages = NULL,
  jobs_title = NULL,
  job_names = NULL,
  collapse_results = FALSE,
  queue = NULL,
  memory = NULL,
  threads = NULL,
  io_saturation = NULL,
  num_proc = NULL,
  queue_flag = "-q @{queue}",
  memory_flag = "-l mem_free=@{memory}G",
  threads_flag = "-pe threads @{threads}",
  io_saturation_flag = "-l io_saturation=@{io_saturation}",
  num_proc_flag = "-l num_proc=@{num_proc}",
  script = .gpatterns.sg_script
)
}
\arguments{
\item{...}{R commands}

\item{command_list}{list of strings with R commands}

\item{opt.flags}{optional flags for qsub command}

\item{max.jobs}{maximal number of simultaneously submitted jobs}

\item{debug}{if 'TRUE', additional reports are printed}

\item{R}{command that launches R}

\item{packages}{packages to load for each command}

\item{jobs_title}{title of job names. names would have the title followed
by a serial number}

\item{job_names}{vector with the names of the jobs}

\item{collapse_results}{collapse return values of the jobs to a data frame.
if not possible - would return the usual list.}

\item{memory}{memory requirments (would be called using \code{memory_flag})}

\item{threads}{threads requirments (would be called using \code{threads_flag})}

\item{io_saturation}{io_saturation requirments (would be called using \code{io_saturation_flag})}

\item{memory_flag}{flag for memory requirment (formatted as in qq)}

\item{threads_flag}{flag for threads requirment (formatted as in qq)}

\item{io_saturation_flag}{flag for io_saturation requirment (formatted as in qq)}
}
\value{
if collapse_results is TRUE: data frame with the results of all jobs (rbinded).
if collapse_results is FALSE returns the same as: \link[misha]{gcluster.run}
}
\description{
Runs R commands on a cluster that supports SGE
}
\examples{

}
\seealso{
\link[misha]{gcluster.run}
}
