% Generated by roxygen2: do not edit by hand
% Please edit documentation in common/gutils.r
\name{gintervals.neighbors1}
\alias{gintervals.neighbors1}
\title{Wrapper around gintervals.neighbours that does not return conflicting colnames}
\usage{
gintervals.neighbors1(intervals1 = NULL, intervals2 = NULL, suffix = "1",
  maxneighbors = 1, mindist = -1e+09, maxdist = 1e+09,
  na.if.notfound = FALSE)
}
\arguments{
\item{suffix}{suffix for conflicting colnames}
}
\description{
Wrapper around gintervals.neighbours that does not return conflicting colnames
}

