% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.apply_tidy_cpgs}
\alias{gpatterns.apply_tidy_cpgs}
\title{Apply a function on tidy cpgs (internally separated by coordinates)}
\usage{
gpatterns.apply_tidy_cpgs(track, f,
  parallel = getOption("gpatterns.parallel"))
}
\arguments{
\item{track}{name of track}

\item{f}{function to apply}

\item{parallel}{execute parallely}
}
\description{
Apply a function on tidy cpgs (internally separated by coordinates)
}

