% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gpatterns.conv_stats}
\alias{gpatterns.conv_stats}
\title{converstion stats}
\usage{
gpatterns.conv_stats(
  bams,
  paired_end = TRUE,
  bismark = FALSE,
  min_mapq = 30,
  bin = .gpatterns.bam2tidy_cpgs_bin
)
}
\description{
converstion stats
}
