% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.get_tidy_cpgs}
\alias{gpatterns.get_tidy_cpgs}
\title{Get tidy cpgs from a track}
\usage{
gpatterns.get_tidy_cpgs(track, intervals = NULL, only_tcpgs = FALSE)
}
\arguments{
\item{track}{name of track}

\item{intervals}{intervals}

\item{only_tcpgs}{use also tracks that have only tidy_cpgs (without cov,meth,unmeth tracks)}
}
\description{
Get tidy cpgs from a track
}
