% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.global_meth_trend}
\alias{gpatterns.global_meth_trend}
\title{Plot global methylation stratified on other tracks}
\usage{
gpatterns.global_meth_trend(tracks,
  strat_track = .gpatterns.cg_cont_500_track, strat_breaks = seq(0, 0.08, by
  = 0.002), intervals = .gpatterns.genome_cpgs_intervals,
  iterator = .gpatterns.genome_cpgs_intervals, min_cov = NULL,
  min_cgs = NULL, names = NULL, width = 500, height = 260,
  fig_fn = NULL, xlab = strat_track, colors = NULL,
  parallel = getOption("gpatterns.parallel"))
}
\arguments{
\item{tracks}{tracks to plot}

\item{strat_track}{track to stratify average methylation by. default is CG content}

\item{strat_breaks}{breaks to determine the bins of strat_track}

\item{intervals}{genomic scope for which the function is applied}

\item{iterator}{track expression iterator (of both tracks and strat_track)}

\item{min_cov}{minimal coverage of each track}

\item{min_cgs}{minimal number of CpGs per bin}

\item{names}{alternative names for the track}

\item{width}{plot width (if fig_fn is not NULL)}

\item{height}{plot height (if fig_fn is not NULL)}

\item{fig_fn}{output filename for the figure (if NULL, figure would be returned)}

\item{xlab}{label for the x axis}

\item{colors}{custom colors}

\item{parallel}{get trends parallely}
}
\value{
list with trend data frame (under 'trend') and the plot (under 'p')
}
\description{
Plot global methylation stratified on other tracks
}
\examples{

}

