% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.import_from_bam}
\alias{gpatterns.import_from_bam}
\title{Creates a track from bam files.}
\usage{
gpatterns.import_from_bam(bams, workdir, track, description,
  conversion = "CT", paired_end = TRUE, downsample = FALSE, dsn = NULL,
  pat_cov_tracks = c(3, 5, 7), pat_freq_len = 2)
}
\arguments{
\item{bams}{bams}

\item{workdir}{workdir}

\item{track}{track}

\item{description}{description}

\item{conversion}{conversion}

\item{paired_end}{paired_end}

\item{downsample}{downsample}

\item{dsn}{dsn}

\item{pat_cov_tracks}{pat_cov_tracks}

\item{pat_freq_len}{pat_freq_len}
}
\description{
Creates a track from bam files for a specific sample.
Using this method is discouraged: if you have large dataset use the
Snakemake pipeline instead.
}

