% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gpatterns.import_from_bam}
\alias{gpatterns.import_from_bam}
\title{Create a track from bam files.}
\usage{
gpatterns.import_from_bam(bams, workdir = NULL, track = NULL,
  steps = "all", conversion = "ct", paired_end = TRUE,
  nbins = nrow(gintervals.all()), groot = GROOT, use_sge = FALSE,
  max_jobs = 400, parallel = getOption("gpatterns.parallel"), ...)
}
\arguments{
\item{bams}{character vector with path of bam files}

\item{workdir}{directory in which the files would be saved}

\item{steps}{steps of the pipeline to do. Possible options are:
'bam2tidy_cpgs', 'filter_dups', 'bind_tidy_cpgs', 'pileup', 'pat_freq', 'pat_cov'}

\item{conversion}{bisulfite conversion. could be 'ct' or 'ga'}

\item{paired_end}{bam files are paired end, with R1 and R2 interleaved}

\item{...}{gpatterns.import_from_tidy_cpgs parameters}
}
\value{
if 'stats' is one of the steps - data frame with statistics. Otherwise none.
}
\description{
Creates a track from bam files for a specific sample.
Use this methods only for small datasets. For large datasets please use
the Snakemake pipeline.
}

