% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{gpatterns.intervs_to_pat_space}
\alias{gpatterns.intervs_to_pat_space}
\title{Generate pattern space for specific genomic loci}
\usage{
gpatterns.intervs_to_pat_space(tracks, intervals, pat_len, max_span = 500,
  contiguous = TRUE, add_rest = FALSE, min_cov = 1)
}
\arguments{
\item{tracks}{tracks}

\item{intervals}{genomic loci to cover}

\item{pat_len}{pattern length}

\item{max_span}{maximum span to look for CpGs}

\item{contiguous}{generate patterns that do not skip a CpG}

\item{add_rest}{add rest of the genome (using gpatterns.tracks_to_pat_space)}

\item{min_cov}{minimal CpG coverage}
}
\description{
Generate pattern space for specific genomic loci
}

