% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_meth.R
\name{gpatterns.spatial_meth_trend}
\alias{gpatterns.spatial_meth_trend}
\title{Plot spatial trend of methylation around intervals}
\usage{
gpatterns.spatial_meth_trend(tracks, intervals, method = "extract",
  dist_breaks = c(-4000000, -1000000, seq(-1000, 1000, 50), 1000000, 4000000),
  iterator = .gpatterns.genome_cpgs_intervals, min_cov = NULL,
  min_cgs = NULL, names = NULL, include.lowest = TRUE, groups = NULL,
  group_name = NULL, add_vline = TRUE, ncol = 2, nrow = 2,
  width = 500, height = 260, fig_fn = NULL, xlab = "Distance (bp)",
  ylim = c(0, 1), title = "", legend = TRUE, colors = NULL,
  parallel = getOption("gpatterns.parallel"))
}
\arguments{
\item{tracks}{tracks to plot}

\item{intervals}{intervals to plot spatial methylation around.
Can be any intervals set, or one of the following 'special' intervals:
'tss', 'exon', 'utr3', 'intron', 'cgi'}

\item{dist_breaks}{breaks to determine the distances from intervals}

\item{iterator}{track expression iterator (of both tracks and strat_track)}

\item{min_cov}{minimal coverage of each track}

\item{min_cgs}{minimal number of CpGs per bin.}

\item{names}{alternative names for the track}

\item{include.lowest}{if 'TRUE', the lowest value of the range determined by breaks is included}

\item{groups}{a vector the same length of \code{tracks} with group for each track. Each group will on a different facet.}

\item{group_name}{name of the grouping variable (e.g. tumor, sample, patient, experiment)}

\item{add_vline}{add vertical line at 0 coordinate}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{width}{plot width (if fig_fn is not NULL)}

\item{height}{plot height (if fig_fn is not NULL)}

\item{fig_fn}{output filename for the figure (if NULL, figure would be returned)}

\item{xlab}{label for the x axis'}

\item{ylim}{ylim of the plot}

\item{title}{title for the plot}

\item{legend}{add legend}

\item{colors}{custom colors}

\item{parallel}{get trends parallely}
}
\value{
list with trend data frame (under 'trend') and the plot (under 'p')
}
\description{
calculates the distance of \code{tracks} values from \code{intervals}
bins the distances based on \code{dist_breaks}, and then calculates the average
methylation \code{m / (m + um)} in each bin and plots the results.
}

