% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gpatterns.track_stats}
\alias{gpatterns.track_stats}
\title{Get QC statistics for track}
\usage{
gpatterns.track_stats(track, ...)
}
\arguments{
\item{track}{track name (or mutiple track names)}

\item{...}{
  Arguments passed on to \code{\link[=gpatterns.get_pipeline_stats]{gpatterns.get_pipeline_stats}}
  \describe{
    \item{\code{tidy_cpgs_stats_dir}}{directory with tidy_cpgs stats files (cannot be used with multiple tracks)}
    \item{\code{uniq_tidy_cpgs_stats_dir}}{directory with filter_dups stats files (cannot be used with multiple tracks)}
    \item{\code{add_mapping_stats}}{add full mapping statistics (singles, discordant etc.)}
    \item{\code{add_insert_length_stats}}{add average insert length}
  }}
}
\description{
Get QC statistics for track
}
