% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pheatmap1}
\alias{pheatmap1}
\title{wrapper around pheatmap}
\usage{
pheatmap1(pmat, annotation = NULL, annotation_colors = NULL,
  annotation_col = NULL, annotation_row = NULL, id_column = TRUE,
  border_color = NA, clustering_callback = function(hc, ...) {    
  dendsort::dendsort(hc) }, ...)
}
\description{
wrapper around pheatmap
}
